backend/requirements.txt
django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pillow==9.4.0